Proposal for new VitessCluster CRD
This is a proposal for a new format for the VitessCluster CDR. It defines a Cluster, Cell, Keyspace, Shard, and Replica. It is designed to allow a user to explicitly specify every aspect of the cluster, or provide limited information and have defaults to instantiate the cluster. Each component also has the ability to accept a "selector" to split configurations into logical parts and compose them into a final resource.

Signed-off-by: Dan Kozlowski <koz@planetscale.com>